Clarify license is CC0
The repository is lacking a license, we have agreed to place it under
CC0.

Add a /LICENSE file, text comes from creative commons plain text version
published at:
http://creativecommons.org/publicdomain/zero/1.0/legalcode.txt

The doc main page offers a download link to retrieve it.

Tweaked the copyright footer to mention WMF and contributors as well as
the license:

 - Copyright 2014, Platform
 + Copyright 2014, Wikimedia Foundation & contributors. Licensed under
 + CC0.

Bug: T94242
Change-Id: Iae080f90d722dc3241be0b40e29d294c8575ee20